fix(scout-for-lol): isolate Dare SQL branch analysis
Why: Keep the generic Dare SQL compiler below repository complexity limits while preserving conservative branch proofs.\n\nWhat: Extract branch analysis and relational policy helpers, retain root query context during decisive-branch evaluation, and simplify finality validation.\n\nVerification: bun run typecheck; bun x eslint src/reports/duckdb/relational-scoutql.ts src/reports/duckdb/relational-scoutql-target-policy.ts src/betting/dare-sql-v3.ts src/betting/dare-sql-v3-branch-analysis.ts src/betting/dare-sql-v3-finality.ts; bun run jscpd; focused Dare SQL Vitest suite (21 tests).
